moviespythonbolt/requirements.txt
# Core database and framework dependencies
neo4j==5.28.2
flask==3.1.0
typing_extensions==4.15.0

# Machine Learning and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Security and Authentication
werkzeug==2.3.6
bcrypt==4.0.1

# Web and API utilities
requests==2.31.0
flask-cors==4.0.0

# Data processing and utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional: For enhanced ML features
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
python-dotenv